+++ tool_graph/graph_view.py
@@ -159,6 +159,7 @@ class GraphPlot(pg.PlotWidget):
                     time_units=self.current_time_units,
                 )
             )
+        self.plotItem.vb.menu.viewAll.triggered.emit()
 
     def on_remove_locations(self, index, start, end):
         """
@@ -1053,9 +1054,13 @@ class GraphDockWidget(QDockWidget):
             graph_widget.add_objects(result.mLayer, [result.mFeature])
             layers_added.add(layer_id)
 
+        from qgis.core import QgsMessageLog
+        QgsMessageLog.logMessage("------------")
         if layers_added:
             tab_index = self.graphTabWidget.indexOf(graph_widget)
             self.graphTabWidget.setCurrentIndex(tab_index)
+            
+            QgsMessageLog.logMessage("EMITTT")
             graph_widget.graph_plot.plotItem.vb.menu.viewAll.triggered.emit()