+++ CHANGES.rst
@@ -4,7 +4,7 @@ Rana Results Analysis changelog
 3.26.15 (unreleased)
 --------------------
 
-- Nothing changed yet.
+- Graph: Fixed missing autoscale (#4126)
 
 
 3.26.14 (2026-05-06)